Decouple report formatting from report blocks
End result: ability to easily add report blocks for
built-in action fields.

This change introduces a new ReportFormatter abstraction. The aim
is to start reusing action field related blocks so we don't need
to have two different sets of blocks, one set in the action pages
and another in the reports app. The report formatters can be
plugged into the existing blocks generated by metaprogramming.

We now start to reuse blocks from actions in reporting especially
for built-in fields. The end goal is to reduce report-specific
code even further and share as much of common formatter code
as possible, but this will happen in the long run. Also, there
is still no full confidence that any built-in field will work
out of the box so some manual work is still needed to add a
new field to reporting.

Also the graphql node types required are now generated
for all the report blocks.