Create management command migrate_wagtailsvg; MANUAL INTERVENTION NEEDED
Before upgrading a deployment to a version newer than this commit, you
need to run the management command `migrate_wagtailsvg`. This should
create an `AplansImage` instance for each instance of `Svg` (the model
from `wagtailsvg`). After this, it should set an FK to the resulting
`AplansImage` in all instances of `CategoryIcon` or
`CommonCategoryIcon` that used to have an FK to an `Svg` instance.

It is possible, however, that the management command fails to create an
`AplansImage` from an `Svg`. This will happen if the SVG file referenced
in the latter does not exist on disk. This can be the case on
development machines, for example, that might not have a copy of all
media files.

Check the output of the management command to see if `Svg` instances
remain. You must make sure that none remain before running a migration
from a newer commit than this one; otherwise there will be
`CategoryIcon` or `CommonCategoryIcon` instances whose `image` field is
NULL. This will raise an integrity error.

On development machines, you might not care too much about icons for
which you don't have the SVG files anyway, so you may just want to
delete those icons:
```
CategoryIcon.objects.filter(image__isnull=True).delete()
CommonCategoryIcon.objects.filter(image__isnull=True).delete()
```