Implement GraphQL query for resolving multiple draft actions
Includes some heavy database query batching and caching, all of which is needed
to reduce the amount of queries to a bearable amount.

The code contains some assumptions about which fields are currently
queried for the mass queries in the UI. Unfortunately a more general
solution similar to the graphql optimizer hints system is unattainable
at the moment for this use case which is not based on querysets, but
a list of actions populated from serialized data. Hence the heavy
use of caching.