Take action revisions into account when querying related orgs
There needs to be quite a lot of special handling for organizations
that are referenced from within Wagtail Revisions of Actions,
inside the responsible_party key. There is no easy way to create
a simple Django ORM annotation (nor filter) to include those
references when calculating organization querysets or action
counts for organizations.

Without this the UI crashes since it is doing quite involved
processing of the organizations (queried separately), combining them
with the actions' organization references (queried within actions).