Change directory structure of uploaded images and renditions
From now on, uploaded images will be stored in
`media/original_images/<year>-<month>/`, and renditions in
`media/images/<year>-<month>/`.

Existing images can be migrated using the new `migrate_images`
management command. This will move the files on disk and update the
paths in the database. It will only affect the original images, i.e.,
the ones in `media/original_images`. For renditions, it should be safe
to call `AplansRendition.objects.all().delete()` because they will be
regenerated on demand.

It's probably also safe to delete all leftover files in
`media/original_images` and `media/images` because they don't seem to be
referenced anywhere.