If client not found for one recipient send other notifications
Rationale: it is difficult to fully enforce that the client
configuration is completely in sync with the person/user
configuration all of the time. Hence, the probability of
there sometimes being users without clients cannot be
easily ruled out and that shouldn't prevent the sending
of notifications to all the other recipients in a plan.