Update permissions for plan features and notification settings
The old ModelAdmin classes for editing plan features and notification
settings in reality only allowed access to superusers, while the
permission helper `ActivePlanPermissionHelper` they used looked more
like it attempted to give access to plan admins as well. Also the code
of the ModelAdmin classes hinted that there should be functionality open
for other users than superusers, for example by defining `panels` and
`superuser_panels` separately for the edit handler.

After internal discussions it was decided that we should avoid risks and
keep the permissions behaviour same as before. This should be assessed
again in the future.